code/crowdflower/table3-rating/requirements.txt

Flask-WTF
Flask-Bootstrap
